add filled orders table
