+++ src/modules/portfolio/containers/positions.js
@@ -1,13 +1,11 @@
 import { connect } from "react-redux";
 import memoize from "memoizee";
 
-import { selectCurrentTimestamp, selectOrphanOrders } from "src/select-state";
+import { selectCurrentTimestamp } from "src/select-state";
 import Positions from "modules/portfolio/components/positions/positions";
 import getLoginAccountPositions from "modules/positions/selectors/login-account-positions";
-import getOpenOrders from "modules/orders/selectors/open-orders";
 import { loadAccountTrades } from "modules/positions/actions/load-account-trades";
 import { triggerTransactionsExport } from "modules/transactions/actions/trigger-transactions-export";
-import { selectMarket } from "modules/markets/selectors/market";
 import { updateModal } from "modules/modal/actions/update-modal";
 import { MODAL_CLAIM_TRADING_PROCEEDS } from "modules/common-elements/constants";
 
@@ -15,7 +13,7 @@ import { createMarketsStateObject } from "modules/portfolio/helpers/create-marke
 
 const mapStateToProps = state => {
   const positions = getLoginAccountPositions();
- 
+
   // NOTE: for data wiring, this should probably be just done as calls for getting openPosition Markets, getting Reporting Markets, and getting Closed Markets respectively from the node and just passed the expected keys below
   const markets = getPositionsMarkets(positions);
   const marketsCount = markets.length;
@@ -40,10 +38,7 @@ const mapDispatchToProps = dispatch => ({
 });
 
 const getPositionsMarkets = memoize(
-  (positions) =>
-    Array.from(
-      new Set([...positions.markets])
-    ),
+  positions => Array.from(new Set([...positions.markets])),
   { max: 1 }
 );
